+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added custom renderer for `line-too-long` to improve its error snippet to highlight the overflow segment instead of the entire line.
 - Improved error message for error C0305 `trailing-newlines` and inserted a "DELETE" comment for each trailing newline.
 - Added a solution to prevent possible large snippets created by the following errors: pylint error C0305, and pycodestyle errors E303, E304
+- Improved RI checking to warn the user when `NameError` is raised due to an omitted `self.`, and the missing name matches an instance attribute.
 
 ### 💫 New checkers